Exclude `ignore` param when creating SigV4 signatures (#46)
* Remove `ignore` param when creating SigV4 signatures

When sending a request using `opensearch-ruby`, specifying the `ignore` parameter causes it to be included in the query string during signature creation. This leads to a `403` error when executing:

```ruby
client.indices.delete(index: 'test', ignore: 404)
```

The reason is that SigV4 includes the `ignore` parameter in the query string when generating the signature.

In fact, at the lower-level transport implementation, the parameter is removed from params before the request is sent:
https://github.com/opensearch-project/opensearch-ruby/blob/3.4.0/lib/opensearch/transport/transport/base.rb#L285

Thus, this library should also remove the ignore parameter before creating the signature.

Signed-off-by: Issei.M <issei.m7@gmail.com>

* Add test

Signed-off-by: Issei.M <issei.m7@gmail.com>

* Update CHANGELOG.md

Signed-off-by: Issei.M <issei.m7@gmail.com>

* Lock opensearch-ruby to versions < 4.0

Signed-off-by: Issei.M <issei.m7@gmail.com>

* Upgrade GitHub official actions 3 to 4

In particular, `actions/upload-artifact@v3` has been deprecated.

Signed-off-by: Issei.M <issei.m7@gmail.com>

* Add Java setup to ensure required version of JDK is installed on test_unreleased

Signed-off-by: Issei.M <issei.m7@gmail.com>

---------

Signed-off-by: Issei.M <issei.m7@gmail.com>